Circos plot of molecular marker distribution across chromosomes

Circos plot of molecular marker distribution across chromosomes

plot_snp_circos(genome_ranges, genes_ranges, marker_ranges)

Arguments

genome_ranges
A GRanges object with chromosome lengths.

genes_ranges
A GRanges object with genomic coordinates of all genes in the genome.

marker_ranges
A GRanges object with genomic positions of molecular markers.

Value

A ggplot object with a circos plot of molecular marker distribution across chromosomes.

See also

Examples

load(gwas)

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following objects are masked from ‘package:base’:
#> 
#>     expand.grid, I, unname

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Error in load(gwas): bad 'file' argument

load(maize_gr)

#> Error in load(maize_gr): bad 'file' argument

load(chr_length)

#> Error in load(chr_length): bad 'file' argument

gwas_list <- split(gwas, gwas$trait)
genes_ranges <- maize_gr[maize_gr$type == "gene", ]
p <- plot_snp_circos(chr_length, genes_ranges, gwas_list)

#> Warning: ‘<’ not meaningful for factors

#> Warning: ‘<’ not meaningful for factors

#> Error in if (any(df[, end_column] < df[, start_column])) {    stop_wrap("End positions should not be smaller than the start positions.")}: missing value where TRUE/FALSE needed